Split DataHelper god-object into partial-class modules (move-only)
Break the ~1160-line DataHelper into a partial static class across five files:
DataHelper.cs keeps the core (dB conversion, ExtractWindow, MinimumAmplitude,
and the option types), DataHelper.Spectrum.cs holds the frequency-response and
harmonic/THD path, DataHelper.Phase.cs the phase and group-delay family,
DataHelper.Impulse.cs impulse + autocorrelation, and DataHelper.Resampling.cs
the log-resample and smoothing. No csproj change (SDK globs *.cs).

Pure move: no signatures, bodies, const values or call ordering changed. All
shared private helpers stay reachable within the one partial class, so every
call site is untouched. Verified byte-identical (sorted non-blank member lines
of the union equal the original) and the 287 dsp tests pass unchanged.
